+++ orm/src/main/java/org/hibernate/tool/hbmlint/detector/RelationalModelDetector.java
@@ -4,7 +4,7 @@ import java.util.Iterator;
 
 import org.hibernate.mapping.Column;
 import org.hibernate.mapping.Table;
-import org.hibernate.tool.hbmlint.Detector;
+import org.hibernate.tool.internal.export.lint.Detector;
 import org.hibernate.tool.internal.export.lint.IssueCollector;
 
 public abstract class RelationalModelDetector extends Detector {
